requirements.txt
numpy==1.26.2
pandas==2.1.3
Pillow==9.3.0
scikit_learn==1.2.2
torch==1.13.1+cu116
torchvision==0.14.1+cu116
transformers==4.36.0.dev0
